refactor handleSubmit and add push to register form
